+++ kladr/rest.sql
@@ -34,7 +34,7 @@ BEGIN
 
     IF jsonb_typeof(pPayload) = 'array' THEN
 
-      FOR r IN SELECT * FROM jsonb_to_recordset(pPayload) AS x(id numeric, fields jsonb)
+      FOR r IN SELECT * FROM jsonb_to_recordset(pPayload) AS x(id integer, fields jsonb)
       LOOP
         FOR e IN EXECUTE format('SELECT %s FROM api.get_address_tree($1)', JsonbToFields(r.fields, GetColumns('address_tree', 'api'))) USING r.id
         LOOP
@@ -44,7 +44,7 @@ BEGIN
 
     ELSE
 
-      FOR r IN SELECT * FROM jsonb_to_record(pPayload) AS x(id numeric, fields jsonb)
+      FOR r IN SELECT * FROM jsonb_to_record(pPayload) AS x(id integer, fields jsonb)
       LOOP
         FOR e IN EXECUTE format('SELECT %s FROM api.get_address_tree($1)', JsonbToFields(r.fields, GetColumns('address_tree', 'api'))) USING r.id
         LOOP
@@ -80,7 +80,7 @@ BEGIN
       pPayload := '{}';
     END IF;
 
-    FOR r IN SELECT * FROM jsonb_to_record(pPayload) AS x(id numeric, code varchar)
+    FOR r IN SELECT * FROM jsonb_to_record(pPayload) AS x(id integer, code varchar)
     LOOP
       FOR e IN SELECT * FROM api.get_address_tree_history(coalesce(r.id, GetAddressTreeId(r.code)))
       LOOP
